map components put into seperate components and comments resolved
