+++ builder/src/components/modals/select-template-modal.tsx
@@ -2,11 +2,9 @@ import React, { FC } from "react";
 import { useDispatch } from "react-redux";
 import { Dialog } from "@headlessui/react";
 import { CgClose } from "react-icons/cg";
-import { toggleModalType, toggleModal } from "redux/modal/modal.reducers";
+import { toggleModal } from "redux/modal/modal.reducers";
 import {ReactComponent as SearchIcon} from 'assets/svgAsIcons/search-icon.svg'
-import Temp1 from "assets/icons/temp-1.png";
-import Temp2 from "assets/icons/temp-2.png";
-import Temp3 from "assets/icons/temp-3.png";
+import SelectTemplateTemplates from "components/utils/select-template-templates";
 
 const TEMPLATE_CATEGORIES = [
   "ALL",
@@ -17,8 +15,6 @@ const TEMPLATE_CATEGORIES = [
   "SHOP",
   "OTHER",
 ];
-
-const TEMPLATES = [Temp1, Temp2, Temp3, Temp3, Temp1, Temp2];
 
 const SelectTemplateModal: FC = () => {
   const dispatch = useDispatch();
@@ -80,36 +76,7 @@ const SelectTemplateModal: FC = () => {
       </div>
       <hr className="bg-hr h-[2px] w-full mt-6" />
       <div className="w-full bg-lower-template">
-        <div className="grid grid-cols-3 gap-4 px-10 pb-12 pt-7">
-          {TEMPLATES.map((temp, index) => {
-            return (
-              <div
-                onClick={() => dispatch(toggleModalType("single"))}
-                key={index}
-                className="bg-white border border-[#E8EAED] rounded-[16px] p-2 cursor-pointer shadow-template-box"
-              >
-                <img
-                  src={temp}
-                  alt="img_temp"
-                  className="w-full rounded-[16px]"
-                  width={314}
-                  height={200}
-                />
-                <div className="flex justify-between items-center font-bold text-[#000000] mt-4 px-2">
-                  <div className="text-[13px] text-[#14142B] opacity-80 ">
-                    Cryptin Next Gen Template
-                  </div>
-                  <div className="text-[10px] text-[#14142B] py-2 px-3 bg-gray-100 rounded-[28px]">
-                    Crypto
-                  </div>
-                </div>
-                <div className="text-[18px] font-[600] text-[#14142B] mt-2 px-2 pb-1">
-                  $399.00
-                </div>
-              </div>
-            );
-          })}
-        </div>
+        <SelectTemplateTemplates />
       </div>
     </Dialog.Panel>
   );
