+++ builder/src/features/dashboard/side-navbar.tsx
@@ -4,12 +4,14 @@ import elements from "assets/icons/elements.png";
 import templates from "assets/icons/templates.png";
 import logo from "assets/icons/buidlfy.png";
 import "styles/components.css";
+import { useDispatch } from "react-redux";
+import { toggleModal, toggleModalType } from "redux/modal/modal.reducers";
 
 interface ISideNavbar {
   className: string;
   setSideElement: (sideElement: string) => void;
-  showSidebar;
-  hideSettingSidebar;
+  showSidebar: () => void;
+  hideSettingSidebar: () => void;
 }
 
 const SideNavbar: FC<ISideNavbar> = ({
@@ -18,6 +20,8 @@ const SideNavbar: FC<ISideNavbar> = ({
   showSidebar,
   hideSettingSidebar,
 }) => {
+  const dispatch = useDispatch()
+
   const handleSidebar = (selectedSidebarElements: string) => {
     setSideElement(selectedSidebarElements);
   };
@@ -31,9 +35,8 @@ const SideNavbar: FC<ISideNavbar> = ({
       <div className="side-border px-4 pt-[1rem]">
         <div
           onClick={() => {
-            showSidebar(true);
-            handleSidebar(SidebarEnum.TEMPLATES);
-            hideSettingSidebar();
+            dispatch(toggleModal(true))
+            dispatch(toggleModalType('template'))
           }}
           className="cursor-pointer"
         >
@@ -42,15 +45,9 @@ const SideNavbar: FC<ISideNavbar> = ({
           </div>
           <h3 className="side-text mt-1">Templates</h3>
         </div>
-        {/* <div className="mt-8 cursor-pointer">
-          <div className="side-icon px-3.5 py-4 rounded-full mt-5">
-            <img src={pages} />
-          </div>
-          <h3 className="side-text mt-1">Pages</h3>
-        </div> */}
         <div
           onClick={() => {
-            showSidebar(true);
+            showSidebar();
             handleSidebar(SidebarEnum.ELEMENTS);
             hideSettingSidebar();
           }}
@@ -61,13 +58,20 @@ const SideNavbar: FC<ISideNavbar> = ({
           </div>
           <h3 className="side-text mt-1">Elements</h3>
         </div>
+        {/* These are commented sidebar elements to be used in future */}
         {/* <div className="mt-8 cursor-pointer">
+          <div className="side-icon px-3.5 py-4 rounded-full mt-5">
+            <img src={pages} />
+          </div>
+          <h3 className="side-text mt-1">Pages</h3>
+        </div> 
+        <div className="mt-8 cursor-pointer">
           <div className="side-icon px-3.5 py-4 rounded-full mt-5">
             <img src={media} />
           </div>
           <h3 className="side-text mt-1">Media</h3>
-        </div> */}
-        {/* <div
+        </div> 
+        <div
           onClick={() => {
             showSidebar(true);
             handleSidebar(SidebarEnum.STYLES);
@@ -79,8 +83,8 @@ const SideNavbar: FC<ISideNavbar> = ({
             <img src={styles} alt="Styles" />
           </div>
           <h3 className="side-text mt-1">Styles</h3>
-        </div> */}
-        {/* <div className="mt-8 cursor-pointer">
+        </div> 
+        <div className="mt-8 cursor-pointer">
           <div className="side-icon px-3.5 py-4 rounded-full mt-5">
             <img src={help} />
           </div>
@@ -93,12 +97,6 @@ const SideNavbar: FC<ISideNavbar> = ({
           <h3 className="side-text mt-1">Settings</h3>
         </div> */}
       </div>
-
-      {/* <Link to="/templates" className="hover:text-black">
-        <div className="mx-6 px-4 py-3 mt-10 rounded-xl hover:bg-blue-100">
-          Templates
-        </div>
-      </Link> */}
     </main>
   );
 };
